FIXED: can add multiple entities of the same type.
Issue was the sorting of  the entity queue in RenderingSystem. If there is only one entity of a certain type/setup, everything is fine because nothing needs to be sorted. Yet, once multiple entities are added the lack of a comparator causes a NullPointerException. I created a new Anonymous comparator that sorts based on position.z values.